+++ custom-elements-manifest.config.mjs
@@ -1,7 +1,0 @@
-const config = {
-  globs: ['src/**/*.js'],
-  exclude: ['src/**/*.css.js', 'src/**/*.stories.js'],
-  litelement: true,
-}
-
-export default config